feat: addLiquidity
you want to do something to batch
all this in the future
if you want addLiquidity to be faster.

also for the showcase you'll want to have
a script that does the createMarkets, feeds
the createMarkets into the addLiquidity calls.

there's an add-liquidity utility
for anyone to enjoy now in the command line.